Fix lots of compiler warnings. Sadly there are two classes that I can't think of a good way to fix right now:
1. A warning about parameter sizes of floats, even when parameter and
   prototype match; this is because of architectures like i386 that
   can't actually pass a float (but it still works, of course).

2. More seriously (because the warnings are more numerous) similar
   warnings about read and write functions. These warnings are mostly
   bogus; I believe they arise for similar reasons (inability of
   calling conventions to deal with integral types smaller than a
   word) but again they work.

If anyone else has ideas on how to reduce the noise without
compromising the code, I'd be grateful.

+++ src/8svx.c
@@ -140,7 +140,7 @@ static int sox_svxstartread(ft_t ft)
                 sox_readdw(ft, &chunksize);
                 if (chunksize & 1)
                         chunksize++;
-                sox_seeki(ft,chunksize,SEEK_CUR);
+                sox_seeki(ft,(int32_t)chunksize,SEEK_CUR);
                 continue;
 
         }
@@ -341,7 +341,7 @@ static void svxwriteheader(ft_t ft, sox_size_t nsamples)
         sox_writedw(ft, nsamples/ft->signal.channels);  /* samples, 1-shot */
         sox_writedw(ft, 0);  /* samples, repeat */
         sox_writedw(ft, 0);  /* samples per repeat cycle */
-        sox_writew(ft, (int) ft->signal.rate); /* samples per second */
+        sox_writew(ft, (uint16_t)ft->signal.rate); /* samples per second */
         sox_writeb(ft,1); /* number of octabes */
         sox_writeb(ft,0); /* data compression (none) */
         sox_writew(ft,1); sox_writew(ft,0); /* volume */
